add a passwd policy generation script
This is analogous to the tpm-policy.sh.  tpm-policy.sh gives the
policy hash for the 'secret' nvindex EA policy.  tpm-policy.passwd.sh
gives it for the 'TPM administrative password" EA policy, since it
also needs to be signed.

Point is we don't want to have to fully provision a TPM using
tpm-setup.sh to generate these files to sign them.

Signed-off-by: Serge Hallyn <serge@hallyn.com>